Refactors state management
.. with the intend to decouple the complexities of setting up
commands and parameters from the maintenance of state (before
the state manager was involved in every setting of parameters,
in order to detect and record the change of state. Now it
just recieves a copy of the complete command after the fact, and
doesn't need to know about command parameters.