Creates read-only command for optimization
When a class retrieves a command, it's 'checked out' from the state
manager, meaning that it is cloned as to not mutate the current state
in case the command has it's parameters modified.

Oftentimes cloning is obsolete however - when the client only wants to
read the values - ie. for log statements or for actual execution.